# back-end/rest_api/src/main/resources/static/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:9115
    description: Generated server url
security:
  - observatoryAuth: []
paths:
  /api/payments/submitproof:
    post:
      tags:
        - payment-controller
      summary: Submit payment proof
      description: Allows a user to submit a payment proof (PDF) along with optional
        metadata to update debts.
      operationId: submitPaymentProof
      parameters:
        - name: toOpName
          in: query
          required: true
          schema:
            type: string
        - name: details
          in: query
          required: false
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - file
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized401Response'
        "500":
          description: Reset failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Generic500Response'
        "400":
          description: Invalid input data or file format
          content:
            application/json:
              schema:
                type: object
        "200":
          description: Payment proof processed successfully
          content:
            application/json:
              schema:
                type: object
...
  /api/payments/toOp:
    get:
      tags:
        - payment-controller
      operationId: paymentsToMe
      responses:
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Unauthorized401Response'
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
  /api/payments/fromOp:
    get:
      tags:
        - payment-controller
      operationId: myPayments
      responses:
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Unauthorized401Response'
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
  /api/payments/debtsto:
    get:
      tags:
        - payment-controller
      operationId: DebtsTo
      responses:
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Unauthorized401Response'
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
  /api/payments/debtsfrom:
    get:
      tags:
        - payment-controller
      operationId: DebtsFrom
      responses:
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Unauthorized401Response'
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
...
components:
  schemas:
    Unauthorized401Response:
      type: object
      properties:
        status:
          type: string
          example: failed
        info:
          type: string
          example: Invalid or expired token
      description: Unauthorized_401_response
    Generic500Response:
      type: object
      properties:
        status:
          type: string
          example: failed
        info:
          type: string
          example: Unexpected internal server error
...
  securitySchemes:
    observatoryAuth:
      type: apiKey
      name: X-OBSERVATORY-AUTH
      in: header